initial depot
This is a heavily modified version of the original and amazing packcc. Changes aim to provide an important performance boost especially by introducing a new management of the memory based on fixed size input and parsing buffers. A bump pointer allocation provides blocks of memory on demand. Consequently, the original code has been cleaned up on all memory freeing code.